Main program now reads from file instead of hard coded matrix.

+++ src/mchi112/hga/CostMatrix.java
@@ -3,7 +3,7 @@ package mchi112.hga;
 public class CostMatrix {
 
 
-    public static void init(int[][] matrix) throws Exception {
+    public static void init(float[][] matrix) throws Exception {
         instance = new CostMatrix(matrix);
     }
 
@@ -12,10 +12,11 @@ public class CostMatrix {
     }
 
     private static CostMatrix instance;
-    private int[][] matrix;
+    private float[][] matrix;
     private int size;
 
-    private CostMatrix(int[][] matrix) throws Exception {
+    private CostMatrix(float[][] matrix) throws Exception {
+
         for (int i = 0; i < matrix.length; i++) {
             if (matrix.length != matrix[i].length) {
                 throw new Exception("Cost matrix must be a square matrix");
